// packages/sdk/src/generated/orval/models/createRoomDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */
import type { CreateRoomDtoFilters } from './createRoomDtoFilters';

export interface CreateRoomDto {
  /** The name of the room */
  name: string;
  /** Filters to apply to the room */
  filters: CreateRoomDtoFilters;
}