// packages/sdk/src/generated/orval/models/roomParticipantStateDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */

export interface RoomParticipantStateDto {
  userId: number;
  userName: string;
  role: string;
  userStatus: string;
  /** Count of liked movie ids this round */
  likedCount?: number;
}