// packages/sdk/src/generated/orval/models/userRoomMembershipDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */
import type { UserRoomMembershipDtoRole } from './userRoomMembershipDtoRole';

export interface UserRoomMembershipDto {
  roomKey: string;
  roomId: string;
  role: UserRoomMembershipDtoRole;
  /** True if this user created the room (room.authorId) */
  isAuthor: boolean;
  userStatus: string;
  matchPhase: string;
  roomStatus: string;
  /** @nullable */
  roomName: string | null;
}